JL8-103 style : Modal & 바텀시트 z-index 때문에 로직 이동
