+++ src/components/animation/ModalAnimation.tsx
@@ -32,7 +32,7 @@ const ModalAnimation: React.FC<ModalAnimationProps> = ({
       {isOpen && (
         <>
           <motion.div
-            className="fixed inset-0 z-40 bg-black bg-opacity-50"
+            className="fixed inset-0 z-50 bg-black bg-opacity-50"
             onClick={onClose}
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
@@ -40,8 +40,7 @@ const ModalAnimation: React.FC<ModalAnimationProps> = ({
           />
           <motion.div
             className={clsx(
-              'fixed left-1/2 top-1/2 z-40 h-auto w-[327px] -translate-x-1/2 -translate-y-1/2 rounded-[32px] bg-white px-6 py-8',
-              'desktop:h-4/5 desktop:w-2/4',
+              'fixed left-1/2 top-1/2 z-50 h-auto w-[327px] -translate-x-1/2 -translate-y-1/2 rounded-[32px] bg-white px-6 py-8',
             )}
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
