+++ src/components/music-play/MusicPlayer.tsx
@@ -24,7 +24,6 @@ const MusicPlayer = () => {
   const { isPlayerOpen, isPlaying, isPlayerModalOpen, stop } =
     useMusicPlayerStore()
   const { musicDetail, url, lyrics, isPending } = usePlayer()
-  console.log('musicDetail', musicDetail)
   const pathname = usePathname()
 
   const stopPlaying =
@@ -91,8 +90,6 @@ const MusicPlayer = () => {
             />
           </div>
           <ActionButtons
-            musicName={musicDetail?.title}
-            artistName={musicDetail?.artist}
             volumeLevel={playerState.volume}
             onVolumeChange={handleVolumeChange}
             className="hidden desktop:flex"
